+++ org.qiweb/org.qiweb.runtime/src/main/java/org/qiweb/runtime/routes/RouteInstance.java
@@ -33,6 +33,7 @@ import org.qiweb.api.http.QueryString;
 import org.qiweb.api.http.RequestHeader;
 import org.qiweb.api.outcomes.Outcome;
 import org.qiweb.api.routes.ControllerParams;
+import org.qiweb.api.routes.ControllerParams.ParamValue;
 import org.qiweb.api.routes.ParameterBinders;
 import org.qiweb.api.routes.Route;
 
@@ -254,7 +255,7 @@ import static org.qiweb.runtime.util.Iterables.toList;
         Map<String, Object> boundParams = new LinkedHashMap<>();
         for( ControllerParams.Param param : controllerParams )
         {
-            if( param.hasForcedValue() )
+            if( ParamValue.FORCED == param.valueKind() )
             {
                 boundParams.put( param.name(), param.forcedValue() );
             }
@@ -277,11 +278,21 @@ import static org.qiweb.runtime.util.Iterables.toList;
                 }
                 if( unboundValue == null )
                 {
-                    throw new ParameterBinderException(
-                        "Parameter named '" + param.name() + "' not found in path nor in query string."
-                    );
+                    if( ParamValue.DEFAULTED == param.valueKind() )
+                    {
+                        boundParams.put( param.name(), param.defaultedValue() );
+                    }
+                    else
+                    {
+                        throw new ParameterBinderException(
+                            "Parameter named '" + param.name() + "' not found in path nor in query string."
+                        );
+                    }
                 }
-                boundParams.put( param.name(), parameterBinders.bind( param.type(), param.name(), unboundValue ) );
+                else
+                {
+                    boundParams.put( param.name(), parameterBinders.bind( param.type(), param.name(), unboundValue ) );
+                }
             }
         }
         return boundParams;
@@ -306,9 +317,20 @@ import static org.qiweb.runtime.util.Iterables.toList;
                     case ":":
                     case "*":
                         ControllerParams.Param controllerParam = controllerParams.get( pathElement.substring( 1 ) );
-                        Object value = controllerParam.hasForcedValue()
-                                       ? controllerParam.forcedValue()
-                                       : parameters.get( controllerParam.name() );
+                        Object value;
+                        if( ParamValue.FORCED == controllerParam.valueKind() )
+                        {
+                            value = controllerParam.forcedValue();
+                        }
+                        else if( ParamValue.DEFAULTED == controllerParam.valueKind()
+                                 && !parameters.containsKey( controllerParam.name() ) )
+                        {
+                            value = controllerParam.defaultedValue();
+                        }
+                        else
+                        {
+                            value = parameters.get( controllerParam.name() );
+                        }
                         unboundPath.append(
                             parameterBinders.unbind(
                                 (Class<Object>) controllerParam.type(),
@@ -342,9 +364,20 @@ import static org.qiweb.runtime.util.Iterables.toList;
             {
                 String qsParam = qsit.next();
                 ControllerParams.Param controllerParam = controllerParams.get( qsParam );
-                Object value = controllerParam.hasForcedValue()
-                               ? controllerParam.forcedValue()
-                               : parameters.get( controllerParam.name() );
+                Object value;
+                if( ParamValue.FORCED == controllerParam.valueKind() )
+                {
+                    value = controllerParam.forcedValue();
+                }
+                else if( ParamValue.DEFAULTED == controllerParam.valueKind()
+                         && !parameters.containsKey( controllerParam.name() ) )
+                {
+                    value = controllerParam.defaultedValue();
+                }
+                else
+                {
+                    value = parameters.get( controllerParam.name() );
+                }
                 unboundPath.append( qsParam ).append( "=" );
                 unboundPath.append(
                     parameterBinders.unbind(
@@ -385,9 +418,15 @@ import static org.qiweb.runtime.util.Iterables.toList;
             {
                 ControllerParams.Param param = it.next();
                 sb.append( param.type().getSimpleName() ).append( " " ).append( param.name() );
-                if( param.hasForcedValue() )
+                switch( param.valueKind() )
                 {
-                    sb.append( " = '" ).append( param.forcedValue() ).append( "'" );
+                    case FORCED:
+                        sb.append( " = '" ).append( param.forcedValue() ).append( "'" );
+                        break;
+                    case DEFAULTED:
+                        sb.append( " ?= '" ).append( param.defaultedValue() ).append( "'" );
+                        break;
+                    default:
                 }
                 if( it.hasNext() )
                 {