+++ org.qiweb/org.qiweb.runtime/src/test/java/org/qiweb/runtime/routes/ParamForcedValueParsingTest.java
@@ -27,6 +27,7 @@ import org.qiweb.runtime.http.QueryStringInstance;
 import static java.util.Collections.singletonList;
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.junit.Assert.assertThat;
+import static org.qiweb.api.routes.RouteBuilder.f;
 import static org.qiweb.api.routes.RouteBuilder.p;
 
 public class ParamForcedValueParsingTest
@@ -39,7 +40,7 @@ public class ParamForcedValueParsingTest
                 new RouteBuilderInstance( app )
                 .route( "GET" )
                 .on( "/foo/:id/bar" )
-                .to( FakeController.class, c -> c.another( p( "id", String.class ), p( "slug", Integer.class, 42 ) ) )
+                .to( FakeController.class, c -> c.another( p( "id", String.class ), f( "slug", Integer.class, 42 ) ) )
                 .modifiedBy( "service", "foo" )
                 .build()
             )
@@ -166,6 +167,56 @@ public class ParamForcedValueParsingTest
                 QueryStringInstance.EMPTY
             );
             assertThat( (String) boundParams.get( "path" ), equalTo( "," ) );
+            assertThat( (Integer) boundParams.get( "num" ), equalTo( 42 ) );
+        }
+        finally
+        {
+            application.passivate();
+        }
+    }
+
+    @Test
+    public void equalsInQuotedForcedValue()
+    {
+        Application application = new ApplicationInstance( new RoutesParserProvider(
+            "GET / com.acme.app.FakeControllerInstance.another( String path = '1=2', Integer num = '42' )"
+        ) );
+        application.activate();
+        try
+        {
+            Route route = application.routes().iterator().next();
+            System.out.println( route );
+            Map<String, Object> boundParams = route.bindParameters(
+                application.parameterBinders(),
+                "/",
+                QueryStringInstance.EMPTY
+            );
+            assertThat( (String) boundParams.get( "path" ), equalTo( "1=2" ) );
+            assertThat( (Integer) boundParams.get( "num" ), equalTo( 42 ) );
+        }
+        finally
+        {
+            application.passivate();
+        }
+    }
+
+    @Test
+    public void questionEqualsInQuotedForcedValue()
+    {
+        Application application = new ApplicationInstance( new RoutesParserProvider(
+            "GET / com.acme.app.FakeControllerInstance.another( String path = '1?=2', Integer num = '42' )"
+        ) );
+        application.activate();
+        try
+        {
+            Route route = application.routes().iterator().next();
+            System.out.println( route );
+            Map<String, Object> boundParams = route.bindParameters(
+                application.parameterBinders(),
+                "/",
+                QueryStringInstance.EMPTY
+            );
+            assertThat( (String) boundParams.get( "path" ), equalTo( "1?=2" ) );
             assertThat( (Integer) boundParams.get( "num" ), equalTo( 42 ) );
         }
         finally